refactor CsvFileTest to import command test
